backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.74",
    "fastapi>=0.121.2",
    "mangum>=0.19.0",
    "openai>=2.8.0",
    "pypdf>=6.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]
